refactor: contract, activity 패키지 이동
- contract : view -> presenter 패키지로 이동
- activity : activity -> view 패키지로 이동
