+++ app/src/main/java/woowacourse/movie/presenter/detail/ReservationDetailContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.presenter.detail
 
 import woowacourse.movie.model.Movie
 import woowacourse.movie.model.Ticket
