+++ app/src/main/java/woowacourse/movie/presenter/finished/ReservationFinishedContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.presenter.finished
 
 import woowacourse.movie.model.Movie
