+++ app/src/main/java/woowacourse/movie/presenter/home/ReservationHomePresenter.kt
@@ -1,7 +1,6 @@
-package woowacourse.movie.presenter
+package woowacourse.movie.presenter.home
 
 import woowacourse.movie.model.Movies
-import woowacourse.movie.view.ReservationHomeContract
 
 class ReservationHomePresenter(
     private val contract: ReservationHomeContract,
