+++ app/src/main/res/layout/activity_reservation_home.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraint_layout_reservation_home"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.ReservationHomeActivity">
+    tools:context=".view.home.ReservationHomeActivity">
 
     <ListView
         android:id="@+id/list_view_reservation_home"
